blog/requirements.txt
Django==2.1.7
django-haystack==2.8.1
django-mdeditor==0.1.13
jieba==0.39
Markdown==2.6.8
mysqlclient==1.4.2.post1
Pillow==5.2.0
PyMySQL==0.9.3
pytz==2018.9
Whoosh==2.7.4
